refactor(read): extract windowed scan driver to scan_stream_windowed.rs (#1143, epic #1116)
Pure mechanical move: relocate run_scan_stream_windowed and drain_scan_window
(impl SSTableReader, added for #1143) out of the over-threshold data_access.rs
into a new sibling submodule scan_stream_windowed.rs. No logic change.

Visibility widened minimally so the moved code compiles from the sibling
module: build_v5_parser and the table_ids_match free function are now
pub(super); the two moved methods are pub(super) so data_access.rs's
scan_stream can still call them via self.

Also reword the #1143 delta_scan.rs comment back to its pre-#1143 line count
to clear the file-size ratchet there (no code change).

data_access.rs: 4187 -> 3988 lines; delta_scan.rs: 3109 -> 3108.
