+++ cqlite-core/src/storage/sstable/reader/delta_scan.rs
@@ -599,9 +599,8 @@ async fn run_scan_delta(
     let reader_arc = std::sync::Arc::clone(&reader);
 
     // The parse closure is synchronous; run it on a blocking thread so it can
-    // use `blocking_send` without stalling the async runtime (the delta-scan
-    // path still materializes via `prepare_delta_scan`; the streaming SELECT
-    // path was moved to a windowed driver in issue #1143).
+    // `blocking_send` without stalling the runtime. Delta-scan still materializes
+    // via `prepare_delta_scan`; streaming SELECT moved to a windowed driver (#1143).
     let parse_result = tokio::task::spawn_blocking(move || -> crate::Result<()> {
         parser.parse_block_emit_delta(
             &stitched,
